controllers: enforce the usage of either --simulation or --device

+++ artiq/frontend/pdq2_controller.py
@@ -27,6 +27,12 @@ def main():
     args = get_argparser().parse_args()
     init_logger(args)
     port = None
+
+    if not args.simulation and args.device is None:
+        raise ValueError("You need to specify either --simulation or "
+                         "-d/--device argument. Use --help for more "
+                         "information.")
+
     if args.simulation:
         port = open(args.dump, "wb")
     dev = Pdq2(url=args.device, dev=port)
